qa: add helm chart tests and readme text
Add tests to the chart.

Helm tests are K8s jobs that Helm can utilize to determine if an
installation was successful. At the moment it is just a simple smoke
test, but in theory a full integration test can be implemented this way
as well.

Add some documentation to the readme. A short howto and explanation for
the options.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>