ci: remove traefik installation from ci pipeline
Remove the traefik installation from the ci pipeline's install test
step. This is unnecessary since the chart no longer depends on traefik.
The ingress installation was never tested during this step, but the
installation of the traefik chart was still necessary since all
dependencies of a chart must be installed regardless of whether they are
used or not.

fixes: #68
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>