ci: fix base ref for release workflow
Change the base ref during version comparison of the release workflow.
This allows changes to happen on main first and then be picked to a
release branch without the linting step failing in the release pipeline.

Otherwise version bumps would need to happend first on the release
branch and the be picked to the main branch, which is error prone and
can get out of sync or be forgotten.

This way all changes happen on the main branch, then the relevant
changes are picked to the release branch. The pipeline then compares the
version in the Chart.yaml in whatever is pushed to the release branch
to the version in the Chart.yaml at the previous HEAD of the release
branch. If that version increased, the release pipeline will continue,
thus avoiding accidental releases by pushing fixes to the release branch
one-by-one.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>