fix: upgrade redeploy strategy
Specify the upgrade redeploy strategy, making sure it recreates the pods
of the deployment rather than rolling out the upgrade.
It is necessary to recreate the pods (i.e. delete the old pods and then
create the new pods) rather than rolling out a new release (creating the
new pods _before_ deleting the old pods), because the PVC is on
ReadWriteOnce mode and thus the volume can only be claimed once by one
pod.

fixes: aquarist-labs/s3gw#397
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>