chart: update logo and source URLs
Update the logo with a URL from the new website.
With the website re-design, the old logo URL is no longer available.
This URL is used in app catalogues like artifacthub.io.

Mention additional repos in the sources list.
This completes the list of source repositories of the software that is
installed when using this chart.

Fixes: aquarist-labs/s3gw#570
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>