chart: validate email for tls issuer
When using the Letsencrypt TLS issuer, a valid email address (the one of
the account requesting the certificate) is required in the .email field
of the values.yaml.
This change introduces a templating check that fails the installation of
the s3gw if the letsencrypt issuer is to be used with the default
(invalid) email, reminding the user to set a valid one.
It also avoids deploying both the self-signed and letsencrypt issuers at
the same time. In case the self-signed issuer is desired it unnecessary
to deploy the letsencrypt issuer as well and with no valid email given
(which isn't necessary for the self signed issuer), the letsencrypt
issued would not deploy successfully.

Fixes: aquarist-labs/s3gw#596
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>